+++ src/ExampleCode/MangoCode.txt
@@ -1,0 +1,9 @@
+// Sample file that will appear in Java++
+
++ class Test
+begin
+    + static void main(String args[])
+    begin
+        println("Hello, this is a print statement in Java++.");
+    end
+end
